Created scripts folder and Label Generation script
I created a scripts folder that will contain all python scripts related to my project. In addition, I have added a Python script called label_generation.py that labels all the data. This Python script ensures labels are in the same file as the numerical data. The rows with a frame count between the Fall_onset_frame and Fall_impact_frame, inclusive, have a label of 1, and the other rows have a label of 0. The sensor data files that are associated with a task that doesn't involve a fall have 0 in every row for the labelling. A 0 means that a fall is not occurring, while a 1 means that a fall is occurring.